Project_Uner/requirements.txt
numpy
torchvision
pillow
natsort
tqdm
easydict
torch==1.9.0
albumentations
tensorboard